refactor: extract shared VCS infrastructure
