+++ package.json
@@ -81,6 +81,7 @@
     "@hunk/session-broker-bun": "workspace:*",
     "@hunk/session-broker-core": "workspace:*",
     "@hunk/term-video": "workspace:*",
+    "@hunk/vcs": "workspace:*",
     "@opentui/core": "^0.5.6",
     "@opentui/react": "^0.5.6",
     "@pierre/diffs": "1.3.5",
