+++ packages/hunk-vcs/src/source.test.ts
@@ -2,7 +2,7 @@ import { afterEach, describe, expect, test } from "bun:test";
 import { mkdtempSync, rmSync, writeFileSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import { readFileTextWithLimit, terminateSourceSubprocess } from "./sourceText";
+import { readFileTextWithLimit, terminateSourceSubprocess } from "./source";
 
 const tempDirs: string[] = [];
