+++ packages/hunk/src/extensions/default/vcs/sapling/commands.ts
@@ -5,9 +5,9 @@ import {
   type ExtensionVcsDiffInput,
   type ExtensionVcsShowInput,
 } from "hunkdiff/extension";
-import { normalizePathForOS } from "../../../../lib/osPath";
-import { describeDiffTargets } from "../diffRange";
-import { runAbortableCommand } from "../asyncProcess";
+import { normalizePathForOS } from "@hunk/vcs/path";
+import { describeDiffTargets } from "@hunk/vcs/diff-target";
+import { runAbortableCommand } from "@hunk/vcs/async-process";
 
 export type SlBackedInput = ExtensionVcsDiffInput | ExtensionVcsShowInput;
