+++ packages/hunk/src/extensions/default/vcs/sapling/index.ts
@@ -11,7 +11,7 @@ import {
   runSlText,
   runSlTextAsync,
 } from "./commands";
-import { describeDiffRange } from "../diffRange";
+import { describeDiffRange } from "@hunk/vcs/diff-target";
 import {
   HUNK_VCS_DETECTION_BASELINE_PRIORITY,
   type ExtensionVcsAdapter,
