+++ tsconfig.json
@@ -18,6 +18,7 @@
       "@hunk/session-broker-bun": ["packages/session-broker-bun/src/index.ts"],
       "@hunk/session-broker-core": ["packages/session-broker-core/src/index.ts"],
       "@hunk/session-broker-node": ["packages/session-broker-node/src/index.ts"],
+      "@hunk/vcs/*": ["packages/hunk-vcs/src/*"],
       "hunkdiff/extension": ["packages/hunk/src/extension-api/index.ts"]
     },
     "strict": true,
